Refactor TrafficCameraApplication to import CameraImages from Overlay directory and add new CameraImage and CameraImages components for improved camera visualization
