+++ nextjs/src/app/_components/Application/TrafficCameraApplication.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 
 import GoogleMap from "~/app/_components/Map/GoogleMap";
-import CameraImages from "~/app/_components/Application/CameraImages";
+import CameraImages from "~/app/_components/Overlay/CameraImages";
 
 import useSocrataData from "~/app/_hooks/useSocrataData";
 import useVisibleCameras from "~/app/_hooks/useVisibleCameras";
