+++ nextjs/src/app/_components/Overlay/CameraImage.tsx
