+++ nextjs/src/app/_components/Overlay/CameraImages.tsx
